% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_gt.R
\name{as_gt.atest}
\alias{as_gt.atest}
\alias{as_gt}
\title{Create a gt table object}
\usage{
\method{as_gt}{atest}(
  x,
  footnote_col = "footnote",
  rowname_col = c(),
  groupname_col = c(),
  simplify = TRUE,
  row_group_as_column = TRUE,
  ...
)

as_gt(x, ...)
}
\arguments{
\item{x}{the object to convert.}

\item{footnote_col}{the column to add footnotes to.}

\item{rowname_col}{column to use as row names, if desired.}

\item{groupname_col}{column to use as groups, if desired.}

\item{simplify}{a logical variable, whether or not to simplify the table first. See Details.}

\item{row_group_as_column}{a logical variable, whether or not to}

\item{...}{additional parameters, passed to \code{gt}.}
}
\description{
Convert an object to a gt table.
}
\details{
Behind the scenes, an \code{atest} object has a number of columns specifying the variables used in the
original formula; these include not only the names of the variables (\code{.y ~ .x | .g})
but also for categorical variables, the values of the variables (\code{.y_value}, \code{.x_value},
\code{.g_value}) and any contrasts (\code{.y_contrast}, \code{.x_contrast}) for two-sample, pairwise,
or paired comparisons. There is also a \code{.terms} variable for the right hand side of a
an \code{atest} about a model.

When \code{simplify = TRUE} (the default), these are converted to more readable columns,
in particular, any values or contrasts are combined with the variable names, and
new variables of \code{response}, \code{variable}, \code{group}, and \code{model} are used instead of
the variables described above.
}
