% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5915_correlation.R
\name{correlation_inference}
\alias{correlation_inference}
\title{Correlation Tests and Confidence Intervals}
\usage{
correlation_inference(
  formula,
  data,
  alternative = c("two.sided", "less", "greater"),
  method = c("pearson", "kendall", "spearman"),
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{formula}{A formula of the form \code{y~x}}

\item{data}{A data frame containing the values in the formula}

\item{alternative}{options}

\item{method}{options}

\item{conf.level}{The desired confidence level for the confidence intervals.}

\item{...}{Additional parameters}
}
\description{
Correlation Tests and Confidence Intervals
}
