% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_statistics.R
\name{descriptive_statistics}
\alias{descriptive_statistics}
\title{Summarize variables}
\usage{
descriptive_statistics(data, ..., digits = 2, compact = TRUE)
}
\arguments{
\item{data}{XX}

\item{...}{XX}

\item{digits}{XX}

\item{compact}{XX}
}
\value{
XX
}
\description{
Summarize variables
}
