% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_plots.R
\name{diag_plots}
\alias{diag_plots}
\title{Display standard diagnostic plots for a model}
\usage{
diag_plots(model)
}
\arguments{
\item{model}{the model to display diagnostic plots for.}
}
\description{
Display standard diagnostic plots for a model.
}
\details{
The four standard diagnostic plots created by \code{plot.lm} are shown, but
using \code{ggplot} graphics techniques. These plots include:
\enumerate{
\item Residual vs Fitted
\item Normal Q-Q
\item Scale-Location
\item Residuals vs Leverage
}
}
