% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pvalue.R
\name{fmt_numbers}
\alias{fmt_numbers}
\title{Format numbers}
\usage{
fmt_numbers(
  data,
  columns = where(is.double) & !any_of(c("p.value", "p.adjust")),
  ...
)
}
\arguments{
\item{data}{XX}

\item{columns}{XX}

\item{...}{XX}
}
\description{
Format numbers
}
