% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pvalue.R
\name{format_pvalue}
\alias{format_pvalue}
\alias{fmt_pvalue}
\title{Format p-values}
\usage{
format_pvalue(
  p,
  digits = 2,
  max.digits = 4,
  justify = TRUE,
  addp = FALSE,
  na = ""
)

fmt_pvalue(data, columns = any_of(c("p.value", "p.adjust")), ...)
}
\arguments{
\item{p}{a vector of p-value(s) to format.}

\item{digits}{the desired number of significant figures.}

\item{max.digits}{the maximum number of decimal places.}

\item{justify}{logical specifying whether or not to align by decimal point.}

\item{addp}{logical specifying whether \verb{p =} or \verb{p <} should be added to the output.}

\item{na}{value to replace missing values with (defaults to a blank).}

\item{data}{data set with columns to format.}

\item{columns}{desired columns to format.}

\item{...}{additional parameters, sent to format_pvalue}
}
\description{
Format p-values
}
