% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_digits.R
\name{format_signif}
\alias{format_signif}
\title{Format to significant digits}
\usage{
format_signif(x, digits = 3, max_small = 6, max_big = 6, keep_big = digits)
}
\arguments{
\item{x}{the numbers to format}

\item{digits}{XX}

\item{max_small}{XX}

\item{max_big}{XX}

\item{keep_big}{XX}
}
\description{
Format to significant digits
}
