% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bars.R
\name{geom_bars}
\alias{geom_bars}
\title{Title}
\usage{
geom_bars(
  mapping = NULL,
  data = NULL,
  stat = "count",
  position = position_dodge2(preserve = "single"),
  ...,
  just = 0.5,
  width = NULL,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{XX}

\item{data}{XX}

\item{stat}{XX}

\item{position}{XX}

\item{...}{XX}

\item{just}{XX}

\item{width}{XX}

\item{na.rm}{XX}

\item{orientation}{XX}

\item{show.legend}{XX}

\item{inherit.aes}{XX}
}
\value{
XX
}
\description{
Title
}
