% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_smooth_binomial.R
\name{geom_smooth_binomial}
\alias{geom_smooth_binomial}
\title{Add a smooth from a logistic regression}
\usage{
geom_smooth_binomial(...)
}
\arguments{
\item{...}{XXX}
}
\description{
Add a smooth from a logistic regression
}
