% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atest.R
\name{gt}
\alias{gt}
\alias{gt.default}
\alias{gt.atest}
\title{Create a gt table object}
\usage{
gt(data, ...)

\method{gt}{default}(data, ...)

\method{gt}{atest}(
  data,
  footnote_col = "footnote",
  rowname_col = "group",
  row_group.sep = " - ",
  ...
)
}
\arguments{
\item{data}{XX}

\item{...}{XX}

\item{footnote_col}{XX}

\item{rowname_col}{XX}

\item{row_group.sep}{XX}
}
\description{
Create a gt table object
}
