% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_cor.R
\name{guess_cor}
\alias{guess_cor}
\title{Create a random scatterplot for guessing correlations}
\usage{
guess_cor(
  seed = NA,
  nonp = FALSE,
  skew = nonp,
  rank = FALSE,
  answer = c("interactive", "show", "hide"),
  return.data = FALSE
)
}
\arguments{
\item{seed}{the random seed used in creating the data.}

\item{nonp}{should non-parametric correlations be shown?}

\item{skew}{should extra skew be randomly added?}

\item{rank}{should the data set be converted to ranks before plotting and computing correlations?}

\item{answer}{option to control interactivity; options are "\code{interactive}", "\code{show}", and "\code{hide}".}

\item{return.data}{normally the plot is returned; if TRUE, the data is returned instead.}
}
\description{
Create a random scatterplot for guessing correlations
}
