% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5915_independence.R
\name{independence_test}
\alias{independence_test}
\alias{independence_test.formula}
\alias{independence_test.default}
\title{Independence Tests for Count Data}
\usage{
independence_test(x, ...)

\method{independence_test}{formula}(formula, data, ...)

\method{independence_test}{default}(
  x,
  method = c("default", "chisq", "exact"),
  correct = FALSE,
  ...
)
}
\arguments{
\item{x}{XX}

\item{...}{XX}

\item{formula}{XX}

\item{data}{XX}

\item{method}{XX}

\item{correct}{XX}
}
\description{
Independence Tests for Count Data
}
