% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5915_nonparametric.R
\name{kruskal_wallis_test}
\alias{kruskal_wallis_test}
\title{Kruskal-Wallis test}
\usage{
kruskal_wallis_test(formula, data)
}
\arguments{
\item{formula}{a formula of the form \code{y ~ x}, where \code{y} is a numeric variable and \code{x} is a factor variable.
To perform test within subgroups, use \code{y ~ x | g},
where \code{x} and \code{g} are factor variables.}

\item{data}{a data frame containing the values in the formula.}
}
\description{
Perform the Kruskal Wallis test to compare two or more groups
}
