% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5915_model.R
\name{model_anova}
\alias{model_anova}
\title{Get model Anova table}
\usage{
model_anova(model, ...)
}
\arguments{
\item{model}{a linear model or generalized linear model.}

\item{...}{additional parameters, passed to \code{car::Anova}.}
}
\description{
Retrieve Anova (analysis of variance) table for a linear model, or analysis
of deviance table for a generalized linear model, as computed by \code{car::Anova}.
}
