% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5915_model.R
\name{model_coefs}
\alias{model_coefs}
\title{Get model coefficients}
\usage{
model_coefs(model, ...)
}
\arguments{
\item{model}{a linear model or generalized linear model.}

\item{...}{additional parameters, passed to \code{broom::tidy}.}
}
\description{
Retrieve coefficients and corresponding confidence intervals and p-values,
as computed by \code{broom::tidy}.
}
