% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5915_model.R
\name{model_glance}
\alias{model_glance}
\title{Get model summary information}
\usage{
model_glance(model, ...)
}
\arguments{
\item{model}{XX}

\item{...}{XX}
}
\value{
XX
}
\description{
Get model summary information
}
