% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5915_model.R
\name{model_means}
\alias{model_means}
\alias{pairwise_model_means}
\alias{model_slopes}
\alias{pairwise_model_slopes}
\title{Get model means and slopes (trends)}
\usage{
model_means(
  model,
  formula,
  cld = TRUE,
  backtransform = TRUE,
  type = if (isTRUE(backtransform)) "response" else "linear",
  ...
)

pairwise_model_means(
  model,
  formula,
  backtransform = TRUE,
  type = if (isTRUE(backtransform)) "response" else "linear",
  ...
)

model_slopes(model, formula, ..., cld = TRUE)

pairwise_model_slopes(model, formula, ...)
}
\arguments{
\item{model}{XX}

\item{formula}{XX}

\item{cld}{XX}

\item{backtransform}{XX}

\item{type}{XX}

\item{...}{XX}
}
\value{
XX
}
\description{
Get model means and slopes (trends)
}
