% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5915_model.R
\name{model_predictions}
\alias{model_predictions}
\title{Get model predictions}
\usage{
model_predictions(
  model,
  at,
  newdata,
  level = 0.95,
  backtransform = TRUE,
  se_fit = FALSE
)
}
\arguments{
\item{model}{a linear model}

\item{at}{the values for each predictor to predict at; uses all combinations of specified values}

\item{newdata}{alternative to at, to specify specific combinations of predictors to predict at}

\item{level}{desired prediction level}

\item{backtransform}{whether or not to backtransform predictions from models on the log scale}

\item{se_fit}{whether or not to report the standard error of the predicted value}
}
\description{
Calculate predictions (fitted values) and prediction intervals for a linear model.
}
