% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtcars2}
\alias{mtcars2}
\title{Motor Trend Car Road Tests}
\format{
A tibble with 32 rows and 12 variables.
\describe{
\item{model}{Make and model of the car}
\item{mpg}{Miles/(US) gallon}
\item{cyl}{Number of cylinders (as a factor)}
\item{disp}{Displacement (cu.in.)}
\item{hp}{Gross horsepower}
\item{drat}{Rear axle ratio}
\item{wt}{Weight (1000 lbs)}
\item{qsec}{1/4 mile time}
\item{vs}{Engine (V-shaped / straight)}
\item{am}{Transmission (automatic / manual)}
\item{gear}{Number of forward gears (as a factor)}
\item{carb}{Number of carburetors (as a factor)}
}
}
\source{
The \code{datasets} package; see \link[datasets:mtcars]{datasets::mtcars}.
}
\usage{
mtcars2
}
\description{
A version of the Motor Trend car road tests data from the datasets package,
but as a tibble with categorical variables stored as factors.
}
\keyword{datasets}
