% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_assignment.R
\name{new_assignment}
\alias{new_assignment}
\title{Create a blank new assignment}
\usage{
new_assignment(
  filename,
  title,
  author,
  open = is_interactive() & isAvailable(),
  template = "template.qmd"
)
}
\arguments{
\item{filename}{XX}

\item{title}{XX}

\item{author}{XX}

\item{open}{XX}

\item{template}{XX}
}
\value{
XX
}
\description{
Create a blank new assignment
}
