% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5915_proportion1.R
\name{one_proportion_test}
\alias{one_proportion_test}
\alias{one_proportion_test.formula}
\alias{one_proportion_test.default}
\alias{wilson_test}
\alias{binomial_test}
\title{One proportion test}
\usage{
one_proportion_test(x, ...)

\method{one_proportion_test}{formula}(formula, data, success, all_success = FALSE, ...)

\method{one_proportion_test}{default}(
  x,
  n,
  null,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  correct = FALSE,
  method = c("default", "wilson", "exact"),
  ...
)

wilson_test(
  x,
  n,
  null,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  correct = FALSE
)

binomial_test(
  x,
  n,
  null,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95
)
}
\arguments{
\item{x}{number of successes}

\item{...}{more things}

\item{formula}{a formula of the form \code{~ x} or \code{x ~ 1}, where \code{x} is a factor variable.
If not a factor, it will be automatically converted.}

\item{data}{a data frame containing the values in the formula}

\item{success}{optional: the level(s) for which proportions should be reported}

\item{all_success}{if TRUE, then proportions for all levels are reported}

\item{n}{number of trials}

\item{null}{null proportion (optional)}

\item{alternative}{character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}

\item{conf.level}{confidence level of the returned confidence interval. Must be a single number between 0 and 1.}

\item{correct}{a logical indicating whether Yates' continuity correction should be applied where possible.}

\item{method}{character string specifying which method to use. One of "default", "wilson", or "exact".}
}
\value{
An atest
}
\description{
A short description XX more ...
}
\details{
More details
}
