% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5915_mean.R
\name{one_t_test}
\alias{one_t_test}
\title{One-sample t-test}
\usage{
one_t_test(
  formula,
  data,
  alternative = c("two.sided", "less", "greater"),
  null,
  conf.level = 0.95,
  backtransform = TRUE,
  var.equal = TRUE
)
}
\arguments{
\item{formula}{XX}

\item{data}{XX}

\item{alternative}{XX}

\item{null}{XX}

\item{conf.level}{XX}

\item{backtransform}{XX}

\item{var.equal}{XX}
}
\description{
One-sample t-test
}
