% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5915_nonparametric.R
\name{one_wilcoxon_inference}
\alias{one_wilcoxon_inference}
\title{One-sample Wilcoxon test}
\usage{
one_wilcoxon_inference(
  formula,
  data,
  alternative = c("two.sided", "less", "greater"),
  null = 0,
  conf.level = 0.95
)
}
\arguments{
\item{formula}{a formula of the form \code{~ y} or \code{y ~ 1}, where \code{y} is a numeric variable.
To perform test within subgroups, use \code{y ~ x} or \code{y ~ 1 | g}, or even \code{y ~ x | g},
where \code{x} and \code{g} are factor variables.}

\item{data}{a data frame containing the values in the formula.}

\item{alternative}{character string specifying the alternative hypothesis, must be one of "\code{two.sided}" (default), "\code{greater}" or "\code{less}".}

\item{null}{a number specifying the null proportion for testing a null hypothesis; if not specified, a null of a center at 0 is used.}

\item{conf.level}{confidence level of the returned confidence interval. Must be a single number between 0 and 1.}
}
\description{
Perform Wilcoxon's signed-rank test on a single sample
}
