% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5915_proportion2.R
\name{paired_proportion_test}
\alias{paired_proportion_test}
\title{Paired proportion test (McNemar's)}
\usage{
paired_proportion_test(
  formula,
  data,
  success,
  alternative = c("two.sided", "less", "greater"),
  correct = FALSE,
  conf.level = NA,
  method = c("default", "wilson", "exact")
)
}
\arguments{
\item{formula}{XX}

\item{data}{XX}

\item{success}{XX}

\item{alternative}{XX}

\item{correct}{XX}

\item{conf.level}{XX}

\item{method}{XX}
}
\description{
Paired proportion test (McNemar's)
}
