% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5915_mean.R
\name{paired_t_inference}
\alias{paired_t_inference}
\title{Paired t-test}
\usage{
paired_t_inference(
  formula,
  data,
  alternative = c("two.sided", "less", "greater"),
  null = 0,
  conf.level = 0.95,
  backtransform = TRUE
)
}
\arguments{
\item{formula}{a formula of the form \code{y2 - y1 ~ 1} or \code{~ y2 -y1}, where \code{y1} and \code{y2} are numeric variables.
To perform test within subgroups, use \code{y2 - y1 ~ x} or \code{y2 - y1 ~ 1 | g}, or even \code{y2- y1 ~ x | g},
where \code{x} and \code{g} are factor variables.}

\item{data}{a data frame containing the values in the formula.}

\item{alternative}{character string specifying the alternative hypothesis, must be one of "\code{two.sided}" (default), "\code{greater}" or "\code{less}".}

\item{null}{a number specifying the null proportion for testing a null hypothesis; if not specified, no hypothesis test is performed.}

\item{conf.level}{confidence level of the returned confidence interval. Must be a single number between 0 and 1.}

\item{backtransform}{if response variable is of form \code{log(y2) - log(y1)}, backtransform
the resulting estimate and confidence interval bounds, so that they
report the ratio of the geometric means rather than the difference on the log scale.}
}
\description{
Compute the mean difference between paired samples, the corresponding
confidence interval, and a p-value for the null hypothesis of equal means
(that is, a difference of zero).
}
