% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5915_mean.R
\name{paired_t_test}
\alias{paired_t_test}
\title{Paired t-test}
\usage{
paired_t_test(
  formula,
  data,
  alternative = c("two.sided", "less", "greater"),
  null = 0,
  conf.level = 0.95,
  backtransform = TRUE
)
}
\arguments{
\item{formula}{XX}

\item{data}{XX}

\item{alternative}{XX}

\item{null}{XX}

\item{conf.level}{XX}

\item{backtransform}{XX}
}
\description{
Paired t-test
}
