% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5915_nonparametric.R
\name{paired_wilcoxon_inference}
\alias{paired_wilcoxon_inference}
\title{Paired Wilcoxon-test}
\usage{
paired_wilcoxon_inference(
  formula,
  data,
  alternative = c("two.sided", "less", "greater"),
  null = 0,
  conf.level = 0.95
)
}
\arguments{
\item{formula}{a formula of the form \code{y2 - y1 ~ 1} or \code{~ y2 -y1}, where \code{y1} and \code{y2} are numeric variables.
To perform test within subgroups, use \code{y2 - y1 ~ x} or \code{y2 - y1 ~ 1 | g}, or even \code{y2- y1 ~ x | g},
where \code{x} and \code{g} are factor variables.}

\item{data}{a data frame containing the values in the formula.}

\item{alternative}{character string specifying the alternative hypothesis, must be one of "\code{two.sided}" (default), "\code{greater}" or "\code{less}".}

\item{null}{a number specifying the null proportion for testing a null hypothesis; if not specified, no hypothesis test is performed.}

\item{conf.level}{confidence level of the returned confidence interval. Must be a single number between 0 and 1.}
}
\description{
Compute the Wilcoxon ranked sum between paired samples.
}
