% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sample_data}
\alias{sample_data}
\alias{sample_lm}
\title{Create a random data set for regression}
\usage{
sample_data(seed, verbose = FALSE)

sample_lm(seed = sample(10000 - 1, 1))
}
\arguments{
\item{seed}{which sample data to use.}

\item{verbose}{logical specifying if messages about sample data should be displayed.}
}
\description{
Create a random data set for regression and display the standard diagnostic plots.
}
