% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_smooth_logistic.R
\name{scale_y_binary}
\alias{scale_y_binary}
\title{Make a binary scale}
\usage{
scale_y_binary(...)
}
\arguments{
\item{...}{Additional parameters, passed to scale_y_continuous}
}
\description{
Make a binary scale
}
