% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_digits.R
\name{set_digits}
\alias{set_digits}
\title{Set desired digits for values in an atest}
\usage{
set_digits(
  x,
  digits = 2,
  decimals,
  rows = seq_len(nrow(x)),
  columns = c("_estimate_", "SE", "conf.low", "conf.high"),
  by = c("SE", "ME"),
  by_row = TRUE
)
}
\arguments{
\item{x}{The atest object}

\item{digits}{The desired number of significant digits.}

\item{decimals}{Alternatively, the number of decimal places to show.}

\item{rows}{The rows to apply this to}

\item{columns}{The columns to apply this to; by default, the columns
corresponding to the estimate, the SE, and the confidence intervals.}

\item{by}{The column that determines the significant digits, by default the SE.}

\item{by_row}{Whether or not to compute the significant digits by row (the default),
or to use the number of digits such that all rows have at least this many.}
}
\description{
Set desired digits for values in an atest
}
