% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim.R
\name{skim}
\alias{skim}
\title{Summarize (skim) the variables in a data set}
\usage{
skim(data, ..., .data_name = NULL)
}
\arguments{
\item{data}{The data frame to skim.}

\item{...}{Columns to select for skimming.}

\item{.data_name}{The name to use for the data. Defaults to the same as data}
}
\description{
\code{skim()}, from the \code{skimr} package, provides a broad overview of a data frame.
It summarizes each type of variable (character, factor, numeric, etc.) in a way
that is appropriate for each. See \link[skimr:skim]{skimr::skim}.
}
