% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_variables.R
\name{summarize_variables}
\alias{summarize_variables}
\title{Summarize variables}
\usage{
summarize_variables(data, ..., digits = 2, compact = TRUE)
}
\arguments{
\item{data}{XX}

\item{...}{XX}

\item{digits}{XX}

\item{compact}{XX}
}
\value{
XX
}
\description{
Summarize variables
}
