% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_functions.R
\name{tab_padding}
\alias{tab_padding}
\alias{tab_compact}
\title{Compact a gt table}
\usage{
tab_padding(x, padding)

tab_compact(x, font.size = 13, padding = 1)
}
\arguments{
\item{x}{XX}

\item{padding}{spacing between rows, in pixels}

\item{font.size}{XX}
}
\value{
XX
}
\description{
Compact a gt table
}
