% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_functions.R
\name{tab_padding}
\alias{tab_padding}
\title{Set gt table padding}
\usage{
tab_padding(x, table.spacing)
}
\arguments{
\item{x}{gt table}

\item{table.spacing}{spacing, in pixels}
}
\value{
a gt.table with modified spacing
}
\description{
Set gt table padding
}
