% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5915_proportion2.R
\name{two_proportion_test.formula}
\alias{two_proportion_test.formula}
\alias{two_proportion_test.default}
\alias{two_proportion_test}
\alias{pairwise_proportion_test}
\title{Two-sample proportion test}
\usage{
\method{two_proportion_test}{formula}(
  formula,
  data,
  success,
  method = c("default", "chisq", "exact"),
  alternative = c("two.sided", "less", "greater"),
  ...
)

\method{two_proportion_test}{default}(
  x,
  n,
  method = c("default", "chisq", "exact"),
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  conf.adjust = 1,
  correct = TRUE,
  ...
)

two_proportion_test(x, ...)

pairwise_proportion_test(
  formula,
  data,
  adjust = c("bonferroni", "holm", "none"),
  reverse = FALSE,
  ...
)
}
\arguments{
\item{formula}{a formula of the form \code{y ~ x}, where \code{y} and \code{x} are both factor variables.
If not factors, they will be automatically converted.
To perform test within subgroups, use \code{y ~ x | g}.}

\item{data}{a data frame containing the values in the formula.}

\item{success}{an optional value specifying the level for which proportions should be reported.}

\item{method}{character string specifying which method to use. One of "\code{default}", "\code{chisq}", or "\code{exact}".}

\item{alternative}{character string specifying the alternative hypothesis, must be one of "\code{two.sided}" (default), "\code{greater}" or "\code{less}".}

\item{...}{further arguments to be passed to submethods, as appropriate.}

\item{x}{vector with count of successes in the two groups}

\item{n}{vector with count of total trials in the two groups}

\item{conf.level}{confidence level of the returned confidence interval. Must be a single number between 0 and 1.}

\item{conf.adjust}{adjust confidence bounds for \code{conf.adjust} simultaneous intervals using the Bonferroni method.
Used internally by \code{pairwise_proportion_test}; should only rarely be used by users.}

\item{correct}{a logical indicating whether Yates' continuity correction should be applied; used for Wilson test only.}

\item{adjust}{method of adjusting p-values for multiple comparisons, one of "\code{bonferroni}", "\code{holm}", or "\code{none}".}

\item{reverse}{reverse the direction of pairwise comparisons.}
}
\value{
A tibble with class \code{atest} containing columns as follows:
\item{difference}{the difference in proportion between the two groups}
\item{conf.low}{lower confidence bound}
\item{conf.high}{upper confidence bound}
\item{chisq.value}{the chi-squared value (if chi-squared method used)}
\item{p.value}{the p-value of the test}
}
\description{
Compute the absolute difference in proportion between two samples, the
corresponding confidence interval, and a p-value for the null hypothesis
of equal proportions (that is, a difference of zero).
}
\details{
Two methods are currently supported, the asymptotic test using the chi-squared
statistic (either with or without continuity correction) or Fisher's exact test.
By default, the chi-squared method is used, with continuity correction determined
as in the \link[stats:prop.test]{stats::prop.test} function, unless the minimum expected count under the
null of equal proportions is less than 5, in which case Fisher's exact test is used.

When Fisher's exact test is used, no confidence interval for the difference
in proportions is reported.

The chi-squared option uses \link[stats:prop.test]{stats::prop.test} and the exact test option uses
\link[stats:fisher.test]{stats::fisher.test}.
}
