% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{two_t_power}
\alias{two_t_power}
\alias{two_t_margin_error}
\title{Power calculations for two sample t tests and confidence intervals}
\usage{
two_t_power(n = NULL, delta = NULL, sd = 1, sig.level = 0.05, power = 0.8)

two_t_margin_error(
  n = NULL,
  me = NULL,
  sd = 1,
  conf.level = 0.95,
  power = 0.8,
  tol = .Machine$double.eps^0.25
)
}
\arguments{
\item{n}{number of observations (per group)}

\item{delta}{true difference between the parameter of interest and the null hypothesis}

\item{sd}{standard deviation (within group)}

\item{sig.level}{significance level}

\item{power}{power of the test, or of having the given margin of error or smaller}

\item{me}{margin of error of the confidence interval}

\item{conf.level}{confidence level}

\item{tol}{numerical tolerance used in root finding, the default providing (at least) four significant digits.}
}
\description{
Power calculations for two sample t tests and confidence intervals
}
