% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5915_mean.R
\name{two_t_test}
\alias{two_t_test}
\alias{pairwise_t_test}
\title{Two Sample t-test}
\usage{
two_t_test(
  formula,
  data,
  alternative = c("two.sided", "less", "greater"),
  null = 0,
  var.equal = FALSE,
  conf.level = 0.95,
  conf.adjust = 1,
  backtransform = TRUE
)

pairwise_t_test(formula, data, ...)
}
\arguments{
\item{formula}{XX}

\item{data}{XX}

\item{alternative}{XX}

\item{null}{XX}

\item{var.equal}{XX}

\item{conf.level}{XX}

\item{conf.adjust}{XX}

\item{backtransform}{XX}

\item{...}{XX}
}
\description{
Two Sample t-test
}
