% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5915_nonparametric.R
\name{two_wilcoxon_inference}
\alias{two_wilcoxon_inference}
\title{Two-sample Wilcoxon test}
\usage{
two_wilcoxon_inference(
  formula,
  data,
  alternative = c("two.sided", "less", "greater"),
  null = 0,
  conf.level = 0.95,
  conf.adjust = 1
)
}
\arguments{
\item{formula}{a formula of the form \code{y ~ x}, where \code{y} is a numeric variable and \code{x} is a factor variable.
To perform test within subgroups, use \code{y ~ x | g},
where \code{x} and \code{g} are factor variables.}

\item{data}{a data frame containing the values in the formula.}

\item{alternative}{character string specifying the alternative hypothesis, must be one of "\code{two.sided}" (default), "\code{greater}" or "\code{less}".}

\item{null}{a number specifying the null proportion for testing a null hypothesis; if not specified, no hypothesis test is performed.}

\item{conf.level}{confidence level of the returned confidence interval. Must be a single number between 0 and 1.}

\item{conf.adjust}{adjust confidence bounds for \code{conf.adjust} simultaneous intervals using the Bonferroni method.
Used internally by \code{pairwise_wilcox_inference}; should only rarely be used by users.}
}
\description{
Perform the Wilcoxon rank sum to compare two groups
}
