% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umncvmstats-package.R
\docType{package}
\name{umncvmstats}
\alias{umncvmstats-package}
\alias{umncvmstats}
\title{umncvmstats ...}
\description{
This package ...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://aaronrendahl.github.io/umncvmstats/}
}

}
\author{
Aaron Rendahl, \email{rend0020@umn.edu}
}
\keyword{internal}
