% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atest.R, R/atest_functions.R
\name{as_gt}
\alias{as_gt}
\alias{as_gt.default}
\alias{as_gt.atest}
\title{Create a gt table object}
\usage{
as_gt(data, ...)

\method{as_gt}{default}(data, ...)

\method{as_gt}{atest}(
  data,
  footnote_col = "footnote",
  rowname_col = ".rowname",
  simplify = TRUE,
  row_group_as_column = TRUE,
  ...
)
}
\arguments{
\item{data}{XX}

\item{...}{XX}

\item{footnote_col}{XX}

\item{rowname_col}{XX}

\item{simplify}{XX}

\item{row_group_as_column}{XX}
}
\description{
Create a gt table object
}
