% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_tests.R
\name{combine_tests}
\alias{combine_tests}
\title{Combine multiple tests together}
\usage{
combine_tests(...)
}
\arguments{
\item{...}{the various tests to be combined.}
}
\description{
Combine multiple tests together
}
