% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pvalue.R
\name{format_pvalue}
\alias{format_pvalue}
\alias{fmt_pvalue}
\alias{fmt_numbers}
\title{Format p-values}
\usage{
format_pvalue(
  p,
  digits = 2,
  max.digits = 4,
  justify = TRUE,
  addp = FALSE,
  na = ""
)

fmt_pvalue(data, columns = any_of(c("p.value", "p.adjust")), ...)

fmt_numbers(
  data,
  columns = where(is.double) & !any_of(c("p.value", "p.adjust")),
  ...
)
}
\arguments{
\item{p}{XX}

\item{digits}{XX}

\item{max.digits}{XX}

\item{justify}{XX}

\item{addp}{XX}

\item{na}{XX}

\item{data}{XX}

\item{columns}{XX}

\item{...}{XX}
}
\value{
XX
}
\description{
Format p-values

Format numbers
}
