% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pvalue.R
\name{fmt_numbers}
\alias{fmt_numbers}
\alias{fmt_pvalue}
\alias{format_pvalue}
\title{Format numbers and p-values}
\usage{
fmt_numbers(
  data,
  columns = where(is.double) & !any_of(c("p.value", "p.adjust")),
  ...
)

fmt_pvalue(data, columns = any_of(c("p.value", "p.adjust")), ...)

format_pvalue(
  p,
  digits = 2,
  max.digits = 4,
  justify = TRUE,
  addp = FALSE,
  na = ""
)
}
\arguments{
\item{data}{data set with columns to format.}

\item{columns}{desired columns to format.}

\item{...}{additional parameters, sent to \link{format_pvalue} (for \code{fmt_pvalue}) or \link{fmt_number} (for \code{fmt_numbers})}

\item{p}{a vector of p-value(s) to format.}

\item{digits}{the desired number of significant figures.}

\item{max.digits}{the maximum number of decimal places.}

\item{justify}{logical specifying whether or not to align by decimal point.}

\item{addp}{logical specifying whether \verb{p =} or \verb{p <} should be added to the output.}

\item{na}{value to replace missing values with (defaults to a blank).}
}
\description{
Format numbers and p-values
}
