% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_smooth_logistic.R
\name{geom_smooth_logistic}
\alias{geom_smooth_logistic}
\title{Add a smooth from a logistic regression}
\usage{
geom_smooth_logistic(...)
}
\arguments{
\item{...}{XXX}
}
\description{
Add a smooth from a logistic regression
}
