% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5915_independence.R
\name{independence_test.formula}
\alias{independence_test.formula}
\alias{independence_test.default}
\alias{independence_test}
\title{Independence Tests for Count Data}
\usage{
\method{independence_test}{formula}(
  formula,
  data,
  method = c("default", "chisq", "exact"),
  correct = FALSE,
  ...
)

\method{independence_test}{default}(
  x,
  method = c("default", "chisq", "exact"),
  correct = FALSE,
  ...
)

independence_test(x, ...)
}
\arguments{
\item{formula}{a formula of the form \code{y ~ x}, where \code{y} and \code{x} are both factor variables.
If not factors, they will be automatically converted.
To perform test within subgroups, use \code{y ~ x | g}.}

\item{data}{a data frame containing the values in the formula.}

\item{method}{character string specifying which method to use. One of "\code{default}", "\code{chisq}", or "\code{exact}".}

\item{correct}{a logical indicating whether Yates' continuity correction should be applied; used for chi-squared test only.}

\item{...}{additional parameters, currently unused.}

\item{x}{instead of a formula, provide the matrix of counts directly.}
}
\description{
Perform an independence test between two categorical variables, using either
the chi-squared method or Fisher's exact method.
}
\details{
By default, the chi-squared method with no continuity correction is used
unless the minimum expected count under the null is less than 5, in which case
Fisher's exact test is used.
}
