% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5915_model.R
\name{model_means}
\alias{model_means}
\alias{pairwise_model_means}
\alias{model_slopes}
\alias{pairwise_model_slopes}
\title{Get model means and slopes (trends)}
\usage{
model_means(model, formula, ..., cld = TRUE)

pairwise_model_means(model, formula, ...)

model_slopes(model, formula, ..., cld = TRUE)

pairwise_model_slopes(model, formula, ...)
}
\arguments{
\item{model}{XX}

\item{formula}{XX}

\item{...}{XX}

\item{cld}{XX}
}
\value{
XX
}
\description{
Get model means and slopes (trends)
}
