% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_assignment.R
\name{new_assignment}
\alias{new_assignment}
\title{Create a blank new assignment}
\usage{
new_assignment(
  filename,
  title,
  author,
  open = is_interactive() & isAvailable()
)
}
\arguments{
\item{filename}{desired filename of new file.}

\item{title}{desired title, to be inserted into new file.}

\item{author}{desired author, to be inserted into new file.}

\item{open}{a logical specifying whether or not to open the new file for editing.}
}
\description{
Create a blank new assignment from the included quarto template.
}
