% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{passfail}
\alias{passfail}
\title{Scores on pre/post test (generated data)}
\format{
A tibble with 50 rows and 4 columns.
\describe{
\item{score1}{Pre-test Score}
\item{score2}{Post-test Score}
\item{pass1}{Pre-test pass/fail status}
\item{pass2}{Post-test pass/fail status}
}
}
\usage{
passfail
}
\description{
Generated pre/post data to demonstrate paired tests.
A score of 80 is considered passing.
}
\keyword{datasets}
