% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_functions.R
\name{tab_compact}
\alias{tab_compact}
\alias{tab_padding}
\title{Compact a gt table}
\usage{
tab_compact(x, font.size = 13, padding = 1)

tab_padding(x, padding)
}
\arguments{
\item{x}{the table to compact.}

\item{font.size}{the desired font size.}

\item{padding}{spacing between rows, in pixels.}
}
\description{
Compact a gt table. Uses \code{tab_options} to set \code{table.font.size} and
the various padding options (see Details).
}
\details{
Padding options that are set are \code{data_row.padding}, \code{summary_row.padding},
\code{grand_summary_row.padding}, \code{footnotes.padding}, \code{source_notes.padding}, and
\code{row_group.padding}; all are set to the same value.
}
