% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_variables.R
\name{tab_compact}
\alias{tab_compact}
\title{Compact a gt table}
\usage{
tab_compact(x, font.size = 13, spacing = 1)
}
\arguments{
\item{x}{XX}

\item{font.size}{XX}

\item{spacing}{XX}
}
\value{
XX
}
\description{
Compact a gt table
}
