% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5915_proportion2.R, R/pairwise.R
\name{two_proportion_test}
\alias{two_proportion_test}
\alias{two_proportion_test.formula}
\alias{two_proportion_test.default}
\alias{pairwise_proportion_test}
\title{Two sample proportion test}
\usage{
two_proportion_test(x, ...)

\method{two_proportion_test}{formula}(
  formula,
  data,
  success,
  method = c("default", "chisq", "exact"),
  alternative = c("two.sided", "less", "greater"),
  ...
)

\method{two_proportion_test}{default}(
  x,
  n,
  method = c("default", "chisq", "exact"),
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  conf.adjust = 1,
  correct = TRUE,
  ...
)

pairwise_proportion_test(formula, data, ...)
}
\arguments{
\item{x}{XX}

\item{...}{XX}

\item{formula}{XX}

\item{data}{XX}

\item{success}{XX}

\item{method}{XX}

\item{alternative}{XX}

\item{n}{XX}

\item{conf.level}{XX}

\item{conf.adjust}{XX}

\item{correct}{XX}
}
\description{
Two sample proportion test
}
